refactor(datatransfer): minor cleanups
add utilties to channel state and extract integration tests
